requirements.txt
transformers
python-magic
tqdm
